Add support for current-bench
The interest of this Pull request, which adds support for current-bench, is to monitor that eqaf operations continue to be resistant to timing attacks during the next Pull requests.

the integration with current-bench requires the installation of the GitHub application : https://github.com/marketplace/ocaml-benchmarks, on this repository, so that the tests are launched automatically after a pull request.

On which ocaml versions would you like to see the tests running?